+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Possibly breaking changes are marked with **BC break**.
 ### Added
 - Connection attributes: add `partnerIP` and `partnerIPv6` (added in SDK 7.50.6, `null` for earlier SDK versions)
 - Add trace level `TRACE_LEVEL_DETAILED` (3), change `TRACE_LEVEL_FULL` to 4 (requires SDK 7.50.9)
+- Add `Connection::setGlobalLogonTimeout()` to set logon timeout
 
 ### Changed
 - Remove `ZEND_ACC_FINAL` from `RemoteFunction` to make it mockable (#133, **BC break**)